Use guard-let to safely unwrap optionals
In the past, the `video` property of `uploadTicket` was assumed to
always have a value, but recently it is nil for some reason. Therefore,
we use guard-let the property to ensure that the app doesn't crash.
Note that the app right now is in the broken state because we don't
have the `video`'s URI yet.