Return empty dictionary
Why do we need to return an empty dictionary instead of `nil`? It's
because we still want to be able to save failures so that the main app
can keep track of failed upload from the share extension. We can only
add new items if we have an empty dictionary.